Refactor snippets to tasks
