+++ foundrytools_cli_2/cli/converter/tasks/variable_to_static.py
