+++ foundrytools_cli_2/cli/print/cli.py
@@ -33,7 +33,7 @@ def print_font_names(
     """
     Prints the name table.
     """
-    from foundrytools_cli_2.cli.print.snippets.print_names import main as print_names
+    from foundrytools_cli_2.cli.print.tasks.print_names import main as print_names
 
     finder = FontFinder(input_path)
     for font in finder.generate_fonts():
