+++ foundrytools_cli_2/cli/ttf/cli.py
@@ -18,7 +18,7 @@ def autohint(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
     Auto-hints the given TrueType fonts using ttfautohint-py.
     """
 
-    from foundrytools_cli_2.cli.ttf.snippets.autohint import ttf_autohint as task
+    from foundrytools_cli_2.cli.ttf.tasks.autohint import ttf_autohint as task
 
     runner = TaskRunner(input_path=input_path, task=task, **options)
     runner.filter.filter_out_ps = True
@@ -32,7 +32,7 @@ def dehint(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
     Removes hinting from the given TrueType fonts.
     """
 
-    from foundrytools_cli_2.cli.ttf.snippets.dehint import ttf_dehint as task
+    from foundrytools_cli_2.cli.ttf.tasks.dehint import ttf_dehint as task
 
     runner = TaskRunner(input_path=input_path, task=task, **options)
     runner.filter.filter_out_ps = True
