+++ foundrytools_cli_2/cli/utils/cli.py
@@ -19,7 +19,7 @@ def font_renamer(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
     """
     Renames the given font files.
     """
-    from foundrytools_cli_2.cli.utils.snippets.font_renamer import main as task
+    from foundrytools_cli_2.cli.utils.tasks.font_renamer import main as task
 
     runner = TaskRunner(input_path=input_path, task=task, **options)
     runner.save_if_modified = False
@@ -37,7 +37,7 @@ def font_organizer(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
     # This is a workaround to make the task work with the current TaskRunner
     options["in_path"] = t.cast(t.Any, input_path)
 
-    from foundrytools_cli_2.cli.utils.snippets.font_organizer import main as task
+    from foundrytools_cli_2.cli.utils.tasks.font_organizer import main as task
 
     runner = TaskRunner(input_path=input_path, task=task, **options)
     runner.save_if_modified = False
@@ -53,6 +53,6 @@ def align_timestamps(input_path: Path, recursive: bool = False) -> None:
     created and modified timestamps stored in their head table.
     """
 
-    from foundrytools_cli_2.cli.utils.snippets.sync_timestamps import main as task
+    from foundrytools_cli_2.cli.utils.tasks.sync_timestamps import main as task
 
     task(input_path, recursive=recursive)
