+++ foundrytools_cli_2/cli/utils/tasks/font_renamer.py
